+++ src/app/AppRegistry.tsx
@@ -43,8 +43,6 @@ import { CitySavedListQueryRenderer } from "./Scenes/City/CitySavedList"
 import { CitySectionListQueryRenderer } from "./Scenes/City/CitySectionList"
 import { CollectionQueryRenderer } from "./Scenes/Collection/Collection"
 import { CollectionFullFeaturedArtistListQueryRenderer } from "./Scenes/Collection/Components/FullFeaturedArtistList"
-import { ConsignmentsHomeScreenQuery } from "./Scenes/Consignments/ConsignmentsHome/ConsignmentsHome"
-import { ConsignmentsSubmissionForm } from "./Scenes/Consignments/ConsignmentsHome/ConsignmentsSubmissionForm"
 import { FairQueryRenderer } from "./Scenes/Fair/Fair"
 import { FairAllFollowedArtistsQueryRenderer } from "./Scenes/Fair/FairAllFollowedArtists"
 import { FairArticlesQueryRenderer } from "./Scenes/Fair/FairArticles"
@@ -107,6 +105,8 @@ import { EditSavedSearchAlertQueryRenderer } from "./Scenes/SavedSearchAlert/Edi
 import { SavedSearchAlertsListQueryRenderer } from "./Scenes/SavedSearchAlertsList/SavedSearchAlertsList"
 import { SearchScreen, SearchScreenQuery } from "./Scenes/Search/Search"
 import { ShowMoreInfoQueryRenderer, ShowQueryRenderer } from "./Scenes/Show"
+import { ConsignmentsHomeScreenQuery } from "./Scenes/SubmitArtwork/ConsignmentsHome/ConsignmentsHome"
+import { ConsignmentsSubmissionForm } from "./Scenes/SubmitArtwork/ConsignmentsHome/ConsignmentsSubmissionForm"
 import { Consignments } from "./Scenes/SubmitArtwork/UploadPhotos/utils"
 import { TagQueryRenderer } from "./Scenes/Tag/Tag"
 import { VanityURLEntityRenderer } from "./Scenes/VanityURL/VanityURLEntity"
