+++ src/app/Scenes/SubmitArtwork/ConsignmentsHome/Components/TextContainer.tsx
