+++ src/app/Scenes/SubmitArtwork/State/SubmissionModel.tsx
@@ -4,11 +4,11 @@ import {
   PhotosFormModel,
 } from "app/Scenes/SubmitArtwork/UploadPhotos/validation"
 import { Action, action } from "easy-peasy"
-import { ConsignmentsSubmissionUtmParams } from "../../Consignments/ConsignmentsHome/ConsignmentsSubmissionForm"
 import {
   artworkDetailsEmptyInitialValues,
   ArtworkDetailsFormModel,
 } from "../ArtworkDetails/validation"
+import { ConsignmentsSubmissionUtmParams } from "../ConsignmentsHome/ConsignmentsSubmissionForm"
 
 export interface ArtworkSubmissionModel {
   submissionId: string
